# Home page: Optimize query for authorized students